Test registry with credentials (#79)
Simplifies the previous version using a single Dockerfile and all test components running in a single container.

In the first iteration of this PR I tried to pack all the code in the Go test. That was a bit awkward in the end, the go test code grew with a lot of exec'ing external tools and doing process management code. I moved most of that now to an script that is embedded into the container that holds and runs the local registry binary and associated tools. The test now could run on its own against an arbitrary registry, although it assumed it needs the registry self-signed certificate to be setup.